+++ src/converter.ts
@@ -1,7 +1,7 @@
 import { Marp, MarpOptions } from '@marp-team/marp-core'
 import { MarpitOptions } from '@marp-team/marpit'
 import * as chromeFinder from 'chrome-launcher/dist/chrome-finder'
-import puppeteer, { PDFOptions } from 'puppeteer-core'
+import puppeteer from 'puppeteer-core'
 import { warn } from './cli'
 import { Engine } from './engine'
 import { error } from './error'
@@ -143,7 +143,7 @@ export class Converter {
           waitUntil: ['domcontentloaded', 'networkidle0'],
         })
 
-        file.buffer = await page.pdf(<PDFOptions>{
+        file.buffer = await page.pdf({
           printBackground: true,
           preferCSSPageSize: true,
         })
@@ -197,8 +197,11 @@ export class Converter {
         ? chromeFinder.wsl
         : chromeFinder[process.platform]
 
+      if (process.env.IS_DOCKER || process.env.CI)
+        args.push('--disable-dev-shm-usage')
+
       if (process.env.IS_DOCKER) {
-        args.push('--no-sandbox', '--disable-dev-shm-usage')
+        args.push('--no-sandbox')
         finder = () => ['/usr/bin/chromium-browser']
       }