+++ test/marp-cli.ts
@@ -9,7 +9,7 @@ import { File } from '../src/file'
 import { Converter, ConvertType } from '../src/converter'
 import { ResolvedEngine } from '../src/engine'
 import { CLIError } from '../src/error'
-import * as preview from '../src/preview'
+import { Preview } from '../src/preview'
 import { Server } from '../src/server'
 import { ThemeSet } from '../src/theme'
 import { Watcher } from '../src/watcher'
@@ -17,14 +17,24 @@ import { Watcher } from '../src/watcher'
 const { version } = require('../package.json')
 const coreVersion = require('@marp-team/marp-core/package.json').version
 const marpitVersion = require('@marp-team/marpit/package.json').version
-const previewMock: any = preview
+const previewEmitter = new EventEmitter()
 
 jest.mock('fs')
 jest.mock('mkdirp')
 jest.mock('../src/preview')
 jest.mock('../src/watcher', () => jest.genMockFromModule('../src/watcher'))
 
-afterEach(() => jest.restoreAllMocks())
+beforeEach(() => {
+  previewEmitter.removeAllListeners()
+  jest
+    .spyOn(Preview.prototype, 'on')
+    .mockImplementation((e, func) => previewEmitter.on(e, func))
+})
+
+afterEach(() => {
+  jest.restoreAllMocks()
+  jest.clearAllMocks()
+})
 
 describe('Marp CLI', () => {
   const assetFn = fn => path.resolve(__dirname, fn)
@@ -116,31 +126,10 @@ describe('Marp CLI', () => {
       })
 
       describe('Preview option', () => {
-        afterEach(() => previewMock.carloOriginal())
-
-        context('when carlo module is loaded (Node >= 7.6.x)', () =>
-          it('outputs help about --preview option', async () => {
-            previewMock.carloMock()
-
-            expect(await run()).toBe(0)
-            expect(error).toBeCalledWith(expect.stringContaining('--preview'))
-            expect(error).toBeCalledWith(
-              expect.not.stringContaining('Requires Node >= 7.6.x')
-            )
-          })
-        )
-
-        context('when carlo module cannot load (Node < 7.6.x)', () =>
-          it('outputs warning of Node version in --preview option', async () => {
-            previewMock.carloUndefined()
-
-            expect(await run()).toBe(0)
-            expect(error).toBeCalledWith(expect.stringContaining('--preview'))
-            expect(error).toBeCalledWith(
-              expect.stringContaining('Requires Node >= 7.6.x')
-            )
-          })
-        )
+        it('outputs help about --preview option', async () => {
+          expect(await run()).toBe(0)
+          expect(error).toBeCalledWith(expect.stringContaining('--preview'))
+        })
 
         context('when CLI is running in an official Docker image', () => {
           beforeEach(() => (process.env.IS_DOCKER = '1'))
@@ -280,60 +269,30 @@ describe('Marp CLI', () => {
       })
 
       context('with --preview option', () => {
-        let open: jest.SpyInstance<preview.ServerPreview['open']>
-
         const run = () =>
           marpCli(['--input-dir', files, '--server', '--preview'])
 
         beforeEach(() => {
           jest.spyOn(cli, 'info').mockImplementation()
           jest.spyOn(Server.prototype, 'start').mockResolvedValue(0)
-
-          open = jest.spyOn(preview.ServerPreview.prototype, 'open')
         })
 
-        afterEach(() => previewMock.carloOriginal())
-
-        context('when carlo module is loaded (Node >= 7.6.x)', () => {
-          let app: EventEmitter & { [func: string]: jest.Mock }
-
-          beforeEach(() => (app = previewMock.carloMock().app))
-
-          it('opens preview window through ServerPreview.open()', async () => {
-            await run()
-            expect(open).toBeCalledTimes(1)
-            expect(app.load).toBeCalledTimes(1)
-
-            const exit = jest.spyOn(process, 'exit').mockImplementation()
-            app.emit('exit')
-            expect(exit).toBeCalled()
-          })
-
-          context('when CLI is running in an official Docker image', () => {
-            beforeEach(() => (process.env.IS_DOCKER = '1'))
-            afterEach(() => delete process.env.IS_DOCKER)
-
-            it('ignores --preview option with warning', async () => {
-              const warn = jest.spyOn(cli, 'warn').mockImplementation()
-
-              await run()
-              expect(open).not.toBeCalled()
-              expect(warn).toBeCalledWith(
-                expect.stringContaining('preview option was ignored')
-              )
-            })
-          })
+        it('opens preview window through Preview.open()', async () => {
+          await run()
+          expect(Preview.prototype.open).toBeCalledTimes(1)
         })
 
-        context('when carlo module cannot load (Node < 7.6)', () => {
+        context('when CLI is running in an official Docker image', () => {
+          beforeEach(() => (process.env.IS_DOCKER = '1'))
+          afterEach(() => delete process.env.IS_DOCKER)
+
           it('ignores --preview option with warning', async () => {
-            previewMock.carloUndefined()
             const warn = jest.spyOn(cli, 'warn').mockImplementation()
 
             await run()
-            expect(open).not.toBeCalled()
+            expect(Preview.prototype.open).not.toBeCalled()
             expect(warn).toBeCalledWith(
-              expect.stringContaining('preview option was ignored')
+              expect.stringContaining('Preview option was ignored')
             )
           })
         })
@@ -649,20 +608,35 @@ describe('Marp CLI', () => {
     })
 
     context('with --preview option', () => {
-      afterEach(() => previewMock.carloOriginal())
+      let warn: jest.SpyInstance<Console['warn']>
 
-      it('outputs warning and starts watching', async () => {
-        const warn = jest.spyOn(cli, 'warn').mockImplementation()
+      beforeEach(() => {
+        warn = jest.spyOn(console, 'warn').mockImplementation()
+        jest.spyOn(process.stdout, 'write').mockImplementation()
+      })
 
-        jest.spyOn(cli, 'info').mockImplementation()
-        previewMock.carloMock()
-        ;(<any>fs).__mockWriteFile()
+      it('opens preview window through Preview.open()', async () => {
+        await marpCli([onePath, '--preview', '-o', '-'])
+        expect(Preview.prototype.open).toBeCalledTimes(1)
 
-        expect(await marpCli([onePath, '--preview'])).toBe(0)
+        // Simualte opening event
+        previewEmitter.emit('opening', '<location>')
         expect(warn).toBeCalledWith(
-          expect.stringContaining('only in server mode')
+          expect.stringContaining('Opening <location>')
         )
-        expect(Watcher.watch).toBeCalledWith([onePath], expect.anything())
+      })
+
+      context('when CLI is running in an official Docker image', () => {
+        beforeEach(() => (process.env.IS_DOCKER = '1'))
+        afterEach(() => delete process.env.IS_DOCKER)
+
+        it('ignores --preview option with warning', async () => {
+          await marpCli([onePath, '--preview', '-o', '-'])
+          expect(Preview.prototype.open).not.toBeCalled()
+          expect(warn).toBeCalledWith(
+            expect.stringContaining('Preview option was ignored')
+          )
+        })
       })
     })
   })
@@ -696,6 +670,18 @@ describe('Marp CLI', () => {
 
         expect(converter.options.inputDir).toBe(assetFn('_files'))
       })
+
+      context('with --preview option', () => {
+        it('opens served address through Preview.open()', async () => {
+          jest.spyOn(console, 'warn').mockImplementation()
+
+          await marpCli(['--server', assetFn('_files'), '--preview'])
+          expect(Preview.prototype.open).toBeCalledTimes(1)
+          expect(Preview.prototype.open).toBeCalledWith(
+            expect.stringMatching(/^http:\/\/localhost:/)
+          )
+        })
+      })
     })
   })
 
@@ -710,6 +696,16 @@ describe('Marp CLI', () => {
         expect(error).toHaveBeenCalledWith(
           expect.stringContaining('specify just one directory')
         )
+      })
+    })
+
+    context('with --preview option', () => {
+      it('opens 2 preview windows through Preview.open()', async () => {
+        jest.spyOn(console, 'warn').mockImplementation()
+        jest.spyOn(process.stdout, 'write').mockImplementation()
+
+        await marpCli([...baseArgs, '--preview', '-o', '-'])
+        expect(Preview.prototype.open).toBeCalledTimes(2)
       })
     })
   })